:truck: Move UnknownValueError to new error module
Refactored UnknownValueError from entry/options.rs into a dedicated error.rs module for better code organization. Updated imports and re-exports to reflect this change.
